Style the issue types mobile selector
