+++ app/views/components/sidebars/filters.js
@@ -1,12 +1,82 @@
+import _ from 'lodash';
 import React from 'react/addons';
 import SidebarFilters from '../filters/sidebar-filters'
+import FiltersActions from '../../../actions/filter-actions';
 
 let FiltersSidebar = React.createClass({
-  
+
+  propTypes: {
+    user: React.PropTypes.object.isRequired,
+    side: React.PropTypes.string.isRequired
+  },
+
+  getInitialState() {
+    return {
+      issueControls: {}
+    }
+  },
+
+  toggleControlState(controls, value) {
+    controls[value] = (controls[value]) ? false : true;
+    this.setState(controls);
+  },
+
+  activeTypes() {
+    return _.chain(this.state.issueControls).map((val, key) => {
+      if(val) {
+        return key
+        }
+      })
+      .compact()
+      .value();
+  },
+
+  /*
+    FIELD: 'type'
+    CRITERIA:  ["defect", "test", "task", "story"]
+  */
+  addItemType(type) {
+    if (type === 'all') {
+      type = ["defect", "test", "task", "story"];
+    }
+    this.toggleControlState(this.state.issueControls, type);
+
+    FiltersActions.update('type', this.activeTypes());
+  },
+
   issueTypesControl() {
+    let issueTypes = ['story', 'task', 'test', 'defect'];
+
+    let issueTypeButtons = _.map(issueTypes, (type) => {
+      let typeClass = {}
+      typeClass[type] = true;
+
+      let linkClasses = React.addons.classSet(_.extend({
+        "active": this.state.issueControls[type]
+      }, typeClass));
+
+      let colorIndicator = React.addons.classSet({
+        'type-color-indicator': true,
+        "active": this.state.issueControls[type]
+      })
+
+      return (
+        <div className='issue-control'>
+          <a href="#" onClick={_.partial(this.addItemType, type)} className={linkClasses}>{type}</a>
+          <div className={`${colorIndicator} ${type}`}></div>
+        </div>
+      )
+    })
+
     return ([
       <li className="drawer-header">
         <a className='drawer-header' href="#">Issue Types</a>
+      </li>,
+      <li className="drawer-subheader">
+        <div className="issue-types-control">
+          {issueTypeButtons}
+        </div>
+        <a href="#" onClick={_.partial(this.addItemType, 'all')} className='all'>All</a>
       </li>
     ])
   },
@@ -19,12 +89,18 @@ let FiltersSidebar = React.createClass({
     ])
   },
 
-  myItems() {
-    console.log('MY ITEMS CONTROL');
+  mine(ev) {
+    ev.preventDefault();
+
+    FiltersActions.update('assigned_to', this.props.user.id);
   },
 
-  myItemsControl() {
-    return <a className="btn btn-primary" href="#" onClick={this.myItems}></a>
+  mineButton() {
+    return (
+      <li>
+        <a href="#" onClick={this.mine} className="btn tbn-primary">My Items</a>
+      </li>
+    )
   },
 
   buildFilterSideBar() {
@@ -37,12 +113,13 @@ let FiltersSidebar = React.createClass({
     })
 
     let minHeight = { 'min-height': `${window.innerHeight}px` };
+    var mineButton = this.mineButton();
 
     return (
       <div style={minHeight} className={classes}>
         <div>Filters</div>
         <ul className="off-canvas-list">
-          {this.myItemsControl()}
+          {this.mineButton()}
           {this.velocityControl()}
           {this.issueTypesControl()}
           <SidebarFilters />
