+++ public/less/utils/font-weights.less
@@ -7,4 +7,3 @@
 @fontWeight-7:   700;
 @fontWeight-8:   800;
 @fontWeight-9:   900;
-@fontWeight-10: 1000;
